pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "requests-glob"
authors = [{name = "David Shea", email = "reallylongword@gmail.com"}]
license = {text = "Apache 2.0"}
description = "File transport adapter for Requests"
readme = "README.rst"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Plugins",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
]

urls = {Homepage = "https://github.com/dashea/requests-file"}
dependencies = ["requests-file>=3.0.0", "sortedcontainers>=1.0.0", "glob2>=0.1"]
version = "0.2"

[tool.distutils.bdist_wheel]
universal = 1

[tool.setuptools]
py-modules = ["requests_glob"]
include-package-data = false

[tool.setuptools_scm]
